feat: refactored usage of shared services in sever-analyis-manager package
