--- packages/server-analysis-manager/src/config/paths.ts
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2022-2024.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import path from 'node:path';
-import type { Analysis } from '@privateaim/core';
-
-let writableDirPath : string | undefined;
-
-export function getWritableDirPath() {
-    if (typeof writableDirPath !== 'undefined') {
-        return writableDirPath;
-    }
-
-    writableDirPath = path.resolve(`${__dirname}../../../writable`);
-    return writableDirPath;
-}
-
-export function generateTrainMinioBucketName(id: Analysis['id']) {
-    return `trains.${id}`;
-}
