+++ packages/server-analysis-manager/src/config/types.ts
@@ -4,9 +4,9 @@
  * For the full copyright and license information,
  * view the LICENSE file that was distributed with this source code.
  */
+import type { Aggregator, Component } from '@privateaim/server-kit';
 
-export * from './amqp';
-export * from './core';
-export * from './logger';
-export * from './storage';
-export * from './vault';
+export type Config = {
+    aggregators: Aggregator[]
+    components: Component[]
+};
