--- packages/server-analysis-manager/src/core/logger/module.ts
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2022-2024.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import type { Logger } from '@privateaim/server-kit';
-import { createLogger } from '@privateaim/server-kit';
-import { getWritableDirPath } from '../../config';
-
-let logger : undefined | any;
-
-/*
-Levels
-{
-  error: 0,
-  warn: 1,
-  info: 2,
-  http: 3,
-  verbose: 4,
-  debug: 5,
-  silly: 6
-}
- */
-
-export function useLogger() : Logger {
-    if (typeof logger !== 'undefined') {
-        return logger;
-    }
-
-    logger = createLogger({
-        directory: getWritableDirPath(),
-    });
-
-    return logger;
-}
