+++ packages/server-analysis-manager/src/start.ts
@@ -5,8 +5,8 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
+import { useLogger } from '@privateaim/server-kit';
 import { createConfig, useEnv } from './config';
-import { useLogger } from './core';
 import { createHttpServer } from './http';
 
 const config = createConfig();
