+++ packages/server-kit/src/services/authup/index.ts
@@ -5,5 +5,6 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
+export * from './helpers';
 export * from './module';
 export * from './singleton';
